Run integration tests against Docker image (#234)

+++ scripts/tools/test-runner.js
@@ -29,6 +29,14 @@ parser.addArgument(['-vb', '--verbose'], {
   help:   'Increase output verbosity',
   action: 'storeTrue' });
 
+parser.addArgument(['-d', '--docker'], {
+  help:   'Test against a Docker container with dind',
+  action: 'storeTrue' });
+
+parser.addArgument(['-n', '--nodeps'], {
+  help:   'Do not start local dependencies',
+  action: 'storeTrue' });
+
 let args = parser.parseArgs();
 
 // -- Separate auth --
@@ -43,16 +51,21 @@ let userDetails = JSON.parse(file);
 
 // -- Create Thunder object --
 let thunder = new ThunderClient(args.endpoint, auth.application, auth.secret);
-let testCases = new TestCases(thunder, userDetails, args.verbose);
+let testCases = new TestCases(thunder, userDetails, args.verbose, args.docker);
 
 // -- Launch required external services --
-console.log('Launching DynamoDB Local...');
-let dynamoProcess = localDynamo.launch(null, 4567);
+let dynamoProcess;
+let sesProcess;
 
-console.log('Launching SES Local...');
-let sesProcess = spawn('npm', ['run', 'ses'], {
-  cwd: __dirname + '/../'
-});
+if (!args.nodeps) {
+  console.log('Launching DynamoDB Local...');
+  dynamoProcess = localDynamo.launch(null, 4567);
+
+  console.log('Launching SES Local...');
+  sesProcess = spawn('npm', ['run', 'ses'], {
+    cwd: __dirname + '/../'
+  });
+}
 
 // -- Run tests --
 console.log('Running full Thunder test...\n');
@@ -76,15 +89,19 @@ async.waterfall(testCases.testPipeline, (err, result) => {
       console.log('Aborting...');
 
       // Clean up
-      dynamoProcess.kill();
-      sesProcess.kill();
+      if (!args.nodeps) {
+        dynamoProcess.kill();
+        sesProcess.kill();
+      }
 
       throw new Error('There are integration test failures');
     });
   } else {
     // Clean up
-    dynamoProcess.kill();
-    sesProcess.kill();
+    if (!args.nodeps) {
+      dynamoProcess.kill();
+      sesProcess.kill();
+    }
 
     process.exit();
   }
